Fix page refresh bug
The filter equality test in componentWillReceiveProps() failed when
there was no filter attribute passed to the List component because then
the filter value was taken from the URL instead of the filter form. In
that case, the filter value is always "{}", but deserializing it leads
to a new object every time, forcing a setFilter() which resets the page.